Add files via upload
The following is a semester long project consisting of me and 5 other people. We were tasked to design a website that with an implemented business idea. We've designed a car rental website in hopes to reach out to the people that have lost their jobs due to the pandemic and want to start being an online driver but don't have a car to start with.